Split out texture into its own class
